+++ src/board/supported/esp_panel_board_config_supported.h
@@ -92,6 +92,7 @@
         + defined(BOARD_JINGCAI_JC8048W550C) \
         /* Waveshare */ \
         + defined(BOARD_WAVESHARE_ESP32_S3_TOUCH_LCD_1_85) \
+        + defined(BOARD_WAVESHARE_ESP32_S3_TOUCH_LCD_1_85_C) \
         + defined(BOARD_WAVESHARE_ESP32_S3_TOUCH_LCD_2_1) \
         + defined(BOARD_WAVESHARE_ESP32_S3_TOUCH_LCD_4_3) \
         + defined(BOARD_WAVESHARE_ESP32_S3_TOUCH_LCD_4_3_B) \
@@ -100,6 +101,13 @@
         + defined(BOARD_WAVESHARE_ESP32_S3_TOUCH_LCD_7) \
         + defined(BOARD_WAVESHARE_ESP32_P4_NANO) \
         /* Viewe */ \
+        + defined(BOARD_VIEWE_SMARTRING) \
+        + defined(BOARD_VIEWE_UEDX24240013_MD50E) \
+        + defined(BOARD_VIEWE_UEDX46460015_MD50ET) \
+        + defined(BOARD_VIEWE_UEDX48480021_MD80E) \
+        + defined(BOARD_VIEWE_UEDX48480021_MD80E_V2) \
+        + defined(BOARD_VIEWE_UEDX48480021_MD80ET) \
+        + defined(BOARD_VIEWE_UEDX48480028_MD80ET) \
         + defined(BOARD_VIEWE_UEDX24320024E_WB_A) \
         + defined(BOARD_VIEWE_UEDX24320028E_WB_A) \
         + defined(BOARD_VIEWE_UEDX24320035E_WB_A) \
@@ -109,6 +117,7 @@
         + defined(BOARD_VIEWE_UEDX80480043E_WB_A) \
         + defined(BOARD_VIEWE_UEDX80480050E_WB_A) \
         + defined(BOARD_VIEWE_UEDX80480050E_WB_A_2) \
+        + defined(BOARD_VIEWE_UEDX80480050E_AC_A) \
         + defined(BOARD_VIEWE_UEDX80480070E_WB_A) \
         > 1
         #error "Multiple boards enabled! Please check file `esp_panel_board_supported_conf.h` and make sure only one board is enabled."
@@ -160,6 +169,8 @@
     /* Waveshare */
     #elif defined(BOARD_WAVESHARE_ESP32_S3_TOUCH_LCD_1_85)
         #include "waveshare/BOARD_WAVESHARE_ESP32_S3_TOUCH_LCD_1_85.h"
+    #elif defined(BOARD_WAVESHARE_ESP32_S3_TOUCH_LCD_1_85_C)
+        #include "waveshare/BOARD_WAVESHARE_ESP32_S3_TOUCH_LCD_1_85_C.h"
     #elif defined(BOARD_WAVESHARE_ESP32_S3_TOUCH_LCD_2_1)
         #include "waveshare/BOARD_WAVESHARE_ESP32_S3_TOUCH_LCD_2_1.h"
     #elif defined(BOARD_WAVESHARE_ESP32_S3_TOUCH_LCD_4_3)
@@ -175,6 +186,20 @@
     #elif defined(BOARD_WAVESHARE_ESP32_P4_NANO)
         #include "waveshare/BOARD_WAVESHARE_ESP32_P4_NANO.h"
     /* Viewe */
+    #elif defined(BOARD_VIEWE_SMARTRING)
+        #include "viewe/BOARD_VIEWE_SMARTRING.h"
+    #elif defined(BOARD_VIEWE_UEDX24240013_MD50E)
+        #include "viewe/BOARD_VIEWE_UEDX24240013_MD50E.h"
+    #elif defined(BOARD_VIEWE_UEDX46460015_MD50ET)
+        #include "viewe/BOARD_VIEWE_UEDX46460015_MD50ET.h"
+    #elif defined(BOARD_VIEWE_UEDX48480021_MD80E)
+        #include "viewe/BOARD_VIEWE_UEDX48480021_MD80E.h"
+    #elif defined(BOARD_VIEWE_UEDX48480021_MD80E_V2)
+        #include "viewe/BOARD_VIEWE_UEDX48480021_MD80E_V2.h"
+    #elif defined(BOARD_VIEWE_UEDX48480021_MD80ET)
+        #include "viewe/BOARD_VIEWE_UEDX48480021_MD80ET.h"
+    #elif defined(BOARD_VIEWE_UEDX48480028_MD80ET)
+        #include "viewe/BOARD_VIEWE_UEDX48480028_MD80ET.h"
     #elif defined(BOARD_VIEWE_UEDX24320024E_WB_A)
         #include "viewe/BOARD_VIEWE_UEDX24320024E_WB_A.h"
     #elif defined(BOARD_VIEWE_UEDX24320028E_WB_A)
@@ -193,6 +218,8 @@
         #include "viewe/BOARD_VIEWE_UEDX80480050E_WB_A.h"
     #elif defined(BOARD_VIEWE_UEDX80480050E_WB_A_2)
         #include "viewe/BOARD_VIEWE_UEDX80480050E_WB_A_2.h"
+    #elif defined(BOARD_VIEWE_UEDX80480050E_AC_A)
+        #include "viewe/BOARD_VIEWE_UEDX80480050E_AC_A.h"
     #elif defined(BOARD_VIEWE_UEDX80480070E_WB_A)
         #include "viewe/BOARD_VIEWE_UEDX80480070E_WB_A.h"
     #else